refatoração pt 1
