+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/controller/CalculoFolha.java
@@ -1,5 +1,5 @@
 package br.com.proway.senior.controller;
 
-public class Folha { //Testing
+public class CalculoFolha { 
 
 }
