+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/model/ColaboradorFolha.java
@@ -2,17 +2,46 @@ package br.com.proway.senior.model;
 
 import java.util.ArrayList;
 
-public final class ColaboradorFolha extends Pessoa {	
-	
+public final class ColaboradorFolha {
+
 	private double salario;
 	private ArrayList<Folha> totalFolhas = new ArrayList<Folha>();
-	private ArrayList<String> dependentes = new ArrayList<String>();	
+	private ArrayList<String> dependentes = new ArrayList<String>();
 	private Ponto ponto;
-	
+	private int id;
+	private String nome;
+	private String email;
+
 	public ColaboradorFolha(String nome, int id, String email, double salario) {
-		super(nome, id, email);				
-		this.salario = salario;				
-	}	
+		this.nome = nome;
+		this.id = id;
+		this.email = email;
+		this.salario = salario;
+	}
+
+	public int getId() {
+		return id;
+	}
+
+	public void setId(int id) {
+		this.id = id;
+	}
+
+	public String getNome() {
+		return nome;
+	}
+
+	public void setNome(String nome) {
+		this.nome = nome;
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
 
 	public double getSalario() {
 		return salario;
@@ -51,11 +80,11 @@ public final class ColaboradorFolha extends Pessoa {
 	 * Adiciona um dependente do Colaborador em sua lista de dependentes totais
 	 * 
 	 * @param p; Nome do dependente
-	 */	 
+	 */
 	public void addDependentes(String p) {
 		this.dependentes.add(p);
 	}
-	
+
 	public Ponto getPonto() {
 		return ponto;
 	}
@@ -63,5 +92,5 @@ public final class ColaboradorFolha extends Pessoa {
 	public void setPonto(Ponto ponto) {
 		this.ponto = ponto;
 	}
-	
+
 }
